# internal/apidocs/openapi.yaml
openapi: "3.1.0"
info:
  title: VTA Farm API
  version: "1.0.0"
  description: |
    REST API for managing VTA setup sessions with per-user namespace isolation.

    ## Authentication

    All accounts use **passkey-only** authentication. There are no passwords.
    Protected endpoints use **httpOnly cookies** set at login.
    Cookie `vtafarm_user` is required for User endpoints; `vtafarm_admin` for Admin endpoints.

    ### In Scalar / browser
    1. Call the passkey login begin endpoint for your account type
    2. Complete the passkey challenge
    3. The browser stores the cookie automatically — all subsequent requests carry it

    ### curl / non-browser clients
    Use the `token` field from the login response body as a Bearer token:
    ```
    Authorization: Bearer <token>
    ```
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://vtafarm-api.firstperson.dev
    description: Production
...
components:
  securitySchemes:
    CookieAuthUser:
      type: apiKey
      in: cookie
      name: vtafarm_user
    CookieAuthAdmin:
      type: apiKey
      in: cookie
      name: vtafarm_admin
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
paths:
...
  /api/v1/admin/passkeys:
    get:
      summary: List admin passkeys
      description: Returns all passkeys registered by the authenticated admin.
      tags: [Admin — Passkeys]
      security:
        - CookieAuthAdmin: []
      responses:
        "200":
          description: List of passkeys
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                      example: "MacBook Touch ID"
                    created_at:
                      type: string
                      format: date-time
                    last_used_at:
                      type: string
                      format: date-time
                      nullable: true
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/passkeys/{id}:
    delete:
      summary: Delete admin passkey
      description: Removes a specific admin passkey.
      tags: [Admin — Passkeys]
      security:
        - CookieAuthAdmin: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Passkey deleted
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Passkey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Admin — Invitations ─────────────────────────────────────────────────────
...
  /api/v1/user/passkeys:
    get:
      summary: List user passkeys
      description: Returns all passkeys registered by the authenticated user.
      tags: [User — Passkeys]
      security:
        - CookieAuthUser: []
      responses:
        "200":
          description: List of passkeys
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                      example: "MacBook Touch ID"
                    created_at:
                      type: string
                      format: date-time
                    last_used_at:
                      type: string
                      format: date-time
                      nullable: true
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/passkeys/{id}:
    delete:
      summary: Delete user passkey
      description: Removes a specific passkey belonging to the authenticated user.
      tags: [User — Passkeys]
      security:
        - CookieAuthUser: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Passkey deleted
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Passkey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── User ────────────────────────────────────────────────────────────────────
...